requirements.txt
APScheduler==3.6.3
asgiref==3.2.10
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.11.8
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cssutils==1.0.2
Django==3.1.1
django-import-export==2.5.0
django-inlinecss==0.3.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
future==0.18.2
idna==2.10
importlib-metadata==2.0.0
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.4
Pillow==8.0.1
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
PyJWT==1.7.1
pynliner==0.8.0
pyparsing==2.4.7
PyPDF4==1.27.0
pytest==6.1.2
pytest-django==4.1.0
pytz==2020.4
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.2
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.2
zipp==3.4.0